% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calc_sigma_nn}
\alias{calc_sigma_nn}
\title{compute euclidian distance for a given matrix and return sigma (1/dist) for the top n nearest neighbor distances}
\usage{
calc_sigma_nn(x, ncpus = 3, squared = TRUE, nn = 10,
  useMedian = TRUE)
}
\arguments{
\item{mat:}{a data.frame or mat object}

\item{ncpus:}{number of cpus to use}

\item{squared:}{boolean, whether to return the squared distance matrix}
}
\description{
compute euclidian distance for a given matrix and return sigma (1/dist) for the top n nearest neighbor distances
}
