% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{commute.time.wrapper}
\alias{commute.time.wrapper}
\title{compute commute time kernel and perform eigen value decomposition}
\usage{
commute.time.wrapper(A, tol = .Machine$double.eps, use.svd = TRUE)
}
\description{
compute commute time kernel and perform eigen value decomposition
}
