% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{cosineKernel}
\alias{cosineKernel}
\title{cosine kernel}
\usage{
cosineKernel(x, ncpus = 1)
}
\arguments{
\item{x:}{an nXp matrix}

\item{ncpus:}{number of cores to use in the calculation}
}
\value{
matrix
}
\description{
compute cosine kernel for a given matrix
cosine similarity for two vectors a,b is
\eqn{cos(\theta)=\frac{\sum_{i=1}^{N}A_{i}B_{i}}{\sqrt{\sum_{i=1}^{N}A_{i}^{2}}\sqrt{\sum_{i=1}^{N}B_{i}^{2}}}}
}
