% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{diffusion_kernel}
\alias{diffusion_kernel}
\title{for a given adjacency matrix and beta values, compute the diffusion kernel}
\usage{
diffusion_kernel(A, beta = 0)
}
\description{
for a given adjacency matrix and beta values, compute the diffusion kernel
}
