% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{diffusionKernel}
\alias{diffusionKernel}
\title{for a given adjacency matrix and beta values, compute the diffusion kernel}
\usage{
diffusionKernel(A, beta = 0)
}
\description{
for a given adjacency matrix and beta values, compute the diffusion kernel
}
