% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{estimateSigma}
\alias{estimateSigma}
\title{estimate sigma for RBF or laplacian kernel
adopted method: sigest from kernlab R package
use the whole data instead of random samples from data}
\usage{
estimateSigma(x, scaled = FALSE, ncpus = 5)
}
\arguments{
\item{scaled}{: Boolean, whether to scale (recommended) the matrix or not for sigma estimation}

\item{x:}{an nXp data.frame or a matrix}

\item{ncpus:}{use the given number of cpus}
}
\value{
vector
}
\description{
estimate sigma for RBF or laplacian kernel
adopted method: sigest from kernlab R package
use the whole data instead of random samples from data
}
