% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{get_dom_term_eigen}
\alias{get_dom_term_eigen}
\title{estimate the number of clusters based on Honarkhah, M and Caers, J (2010) methodology
source: Honarkhah, M and Caers, J (2010). "Stochastic Simulation of Patterns Using Distance-Based Pattern Modeling".
                                           Mathematical Geosciences 42 (5): 487–517. doi:10.1007/s11004-010-9276-7
pdf source: https://pangea.stanford.edu/ERE/pdf/pereports/PhD/Honarkhah2011.pdf
springer pay page: http://link.springer.com/article/10.1007%2Fs11004-010-9276-7}
\usage{
get_dom_term_eigen(obj)
}
\arguments{
\item{obj:}{a list object, output from 'eigen' function}
}
\value{
data.frame
}
\description{
estimate the number of clusters based on Honarkhah, M and Caers, J (2010) methodology
source: Honarkhah, M and Caers, J (2010). "Stochastic Simulation of Patterns Using Distance-Based Pattern Modeling".
                                           Mathematical Geosciences 42 (5): 487–517. doi:10.1007/s11004-010-9276-7
pdf source: https://pangea.stanford.edu/ERE/pdf/pereports/PhD/Honarkhah2011.pdf
springer pay page: http://link.springer.com/article/10.1007%2Fs11004-010-9276-7
}
