% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf.R
\name{get_tf_binary}
\alias{get_tf_binary}
\title{given a data.frame containing gene and associated transcription factors, return binary matrix
1 if tf binds to gene, 0 if not}
\usage{
get_tf_binary(df)
}
\arguments{
\item{df:}{a data.frame, first column must be gene ids and second column must be transcription factors}
}
\value{
matrix
}
\description{
given a data.frame containing gene and associated transcription factors, return binary matrix
1 if tf binds to gene, 0 if not
}
