% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf.R
\name{get_tf_idf}
\alias{get_tf_idf}
\title{given a data.frame containing gene and associated transcription factors, calculate tf idf values}
\usage{
get_tf_idf(df, notf = FALSE)
}
\arguments{
\item{df:}{a data.frame, first column must be gene ids and second column must be transcription factors}

\item{notf:}{Boolean, if true, only idf value will be calculated}
}
\value{
matrix
}
\description{
given a data.frame containing gene and associated transcription factors, calculate tf idf values
}
