% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_wrapper}
\alias{graph_wrapper}
\title{a wrapper function for rbf_wrapper and graph_laplacian functions}
\usage{
graph_wrapper(x, sigma = 1, lap = FALSE, ncpus = 5,
  type = c("norm", "sym", "rw"))
}
\description{
a wrapper function for rbf_wrapper and graph_laplacian functions
}
