% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{linearKernel}
\alias{linearKernel}
\title{linear kernel}
\usage{
linearKernel(x, ncpus)
}
\arguments{
\item{x:}{a matrix/data.frame object}

\item{cores:}{number of cpus to use}
}
\value{
matrix
}
\description{
compute linear kernel
for a given matrix, compute linear kernel as:
\eqn{k(x,y)=x^{T}y+c}
}
