% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{ma}
\alias{ma}
\title{helper function
calculate moving average}
\usage{
ma(vec, n = 3)
}
\arguments{
\item{vec:}{a vector of values}

\item{n:}{number of values to calculate moving average from}
}
\value{
vector
}
\description{
helper function
calculate moving average
}
