% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{neumannKernel}
\alias{neumannKernel}
\title{Neumann diffusion kernel}
\usage{
neumannKernel(A, frac = 0.9)
}
\arguments{
\item{A}{a graph adjacency matrix}

\item{frac}{a value 0<frac<1 to calculate alpha}
}
\value{
matrix
}
\description{
for a given graph adjacency matrix, calculate von Neumann diffusion kernel
defined as K = inv(I-alpha*A), where alpha is defined as 0<alpha<sig(A)
and sig(A) is max(abs(eigen(A)$values))
}
