% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pinv}
\alias{pinv}
\title{For a given matrix calculate the psuedo inverse}
\usage{
pinv(m, tol = 1e-10)
}
\arguments{
\item{m}{: a square matrix}

\item{tol}{: tolerance level (default: 1e-10) : tolerance level}
}
\value{
matrix
}
\description{
For a given matrix calculate the psuedo inverse
}
