% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{rbf_nn}
\alias{rbf_nn}
\title{compute adaptive rbf or laplacian kernel
adaptive: instead of a golobal gamma value, caluclate gamma per point as the mean/meadian of n nearset neighbours,}
\usage{
rbf_nn(x, vec = NULL, ncpus = 10, lap = FALSE, nn = 20,
  useMedian = TRUE)
}
\description{
compute adaptive rbf or laplacian kernel
adaptive: instead of a golobal gamma value, caluclate gamma per point as the mean/meadian of n nearset neighbours,
}
