% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{rbfKernel}
\alias{rbfKernel}
\title{Radial basis function kernel}
\usage{
rbfKernel(x, sigma = NULL, scaled = FALSE, ncpus = 3, lap = FALSE)
}
\arguments{
\item{scaled}{: Boolean, whether to scale (recommended) the matrix or not}

\item{x:}{an nXp data.frame or a matrix}

\item{sigma:}{if Sigma is NULL, calculate sigma value, else use the value provided}

\item{ncpus:}{use the given number of cpus}

\item{lap:}{Boolean, whether to calculate laplacian kernel or rbf kernel, if FALSE (default) returns rbf kernel}
}
\value{
matrix
}
\description{
wrapper function for rbf/laplacian kernel and sigma.sample
estimate sigma and use the sigma for computing laplacian kernel
}
