% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{rct_wrapper}
\alias{rct_wrapper}
\title{compute regularized commute time kernel and perform eigen value decomposition}
\usage{
rct_wrapper(A, alpha = 0.95)
}
\description{
compute regularized commute time kernel and perform eigen value decomposition
}
