% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{regularized_commute_kernel}
\alias{regularized_commute_kernel}
\title{For a given graph, compute regularized commute time kernel
formula source: "François Fouss, Kevin Francoisse, Luh Yen, Alain Pirotte, and Marco Saerens.
An experimental investigation of kernels on graphs for collaborative recommendation and semisupervised classification.
Neural Netw. 31 (July 2012), 53-72. DOI=10.1016/j.neunet.2012.03.001 http://dx.doi.org/10.1016/j.neunet.2012.03.001."}
\usage{
regularized_commute_kernel(A, alpha = 0.95)
}
\arguments{
\item{A}{: a graph adjacency matrix}

\item{alpha}{: The probability of random walker staying at each step. 1-alpha is the probability of the walker
: "evaporating" at each step.}
}
\value{
matrix
}
\description{
For a given graph, compute regularized commute time kernel
formula source: "François Fouss, Kevin Francoisse, Luh Yen, Alain Pirotte, and Marco Saerens.
An experimental investigation of kernels on graphs for collaborative recommendation and semisupervised classification.
Neural Netw. 31 (July 2012), 53-72. DOI=10.1016/j.neunet.2012.03.001 http://dx.doi.org/10.1016/j.neunet.2012.03.001."
}
