% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf.R
\name{tfIdf}
\alias{tfIdf}
\title{Transcription factor tf idf}
\usage{
tfIdf(df, notf = FALSE)
}
\arguments{
\item{df}{a data.frame, first column must be gene ids and second column must be transcription factors}

\item{notf}{Boolean, if true, only idf value will be calculated}
}
\value{
matrix
}
\description{
Given a data.frame containing gene and associated transcription factors, calculate tf idf
Term frequency, iverse document frequency values
}
