% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{distance}
\alias{distance}
\title{compute euclidian distance for a given matrix}
\usage{
distance(x, ncpus = 3, squared = TRUE)
}
\arguments{
\item{ncpus}{number of cpus to use}

\item{squared}{boolean, whether to return the squared distance matrix}

\item{mat}{a data.frame or mat object}
}
\description{
compute euclidian distance for a given matrix
}
